Backend/app/requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.11.18 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.11" and python_version < "4.0"
bitarray==3.3.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.4.26 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_version < "4.0"
ckzg==2.1.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cytoolz==1.0.1 ; python_version >= "3.11" and python_version < "4.0" and implementation_name == "cpython"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
eth-abi==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
eth-account==0.13.7 ; python_version >= "3.11" and python_version < "4.0"
eth-hash==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
eth-keyfile==0.8.1 ; python_version >= "3.11" and python_version < "4.0"
eth-keys==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
eth-rlp==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
eth-typing==5.2.1 ; python_version >= "3.11" and python_version < "4.0"
eth-utils==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi-cli==0.0.7 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.115.12 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
hexbytes==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.4.3 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
parsimonious==0.10.0 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
py-solc-x==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
pycryptodome==3.22.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.4 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.3.5 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.11" and python_version < "4.0"
pyunormalize==16.0.0 ; python_version >= "3.11" and python_version < "4.0"
pywin32==310 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rich-toolkit==0.14.4 ; python_version >= "3.11" and python_version < "4.0"
rich==14.0.0 ; python_version >= "3.11" and python_version < "4.0"
rlp==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.11" and python_version < "4.0"
toolz==1.0.0 ; python_version >= "3.11" and python_version < "4.0" and (implementation_name == "pypy" or implementation_name == "cpython")
typer==0.15.3 ; python_version >= "3.11" and python_version < "4.0"
types-requests==2.32.0.20250328 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.34.2 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
web3==7.11.0 ; python_version >= "3.11" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.20.0 ; python_version >= "3.11" and python_version < "4.0"
